pkg/sovits/requirements_win.txt
librosa==0.9.1
fairseq==0.12.2
ffmpeg-python
Flask==2.1.2
Flask_Cors==3.0.10
gradio>=3.7.0
numpy
playsound==1.3.0
PyAudio==0.2.12
pydub==0.25.1
pyworld
requests==2.28.1
scipy
sounddevice==0.4.5
SoundFile==0.10.3.post1
starlette==0.19.1
tqdm==4.63.0
rich
loguru
torchcrepe
scikit-maad
praat-parselmouth
onnx
onnxsim
onnxoptimizer
tensorboard
tensorboardX
transformers
edge_tts
langdetect
pyyaml
pynvml
faiss-cpu

